administrative ; resize bug fix from Joern ; warnings-- ; add doc ; buffer overflow bugfix for 71384 ; cleanup. (See changelog entry for details).
svn path=/trunk/kdepim/; revision=304327

+++ kpilot/kpilot/addressWidget.cc
@@ -179,7 +179,7 @@ void AddressWidget::hideComponent()
 	{
 		fAddressList.clear();
 		fListBox->clear();
-	
+
 		updateWidget();
 	}
 }
@@ -215,7 +215,7 @@ void AddressWidget::postHotSync()
 {
 	FUNCTIONSETUP;
 
-	if ( shown ) 
+	if ( shown )
 	{
 		fAddressList.clear();
 		showComponent();
@@ -463,7 +463,7 @@ void AddressWidget::slotCreateNewRecord()
 			<< dbPath()
 			<< "/AddressDB"
 			<< " and got pointer @"
-			<< (int) myDB
+			<< (long) myDB
 			<< " with status "
 			<< ( myDB ? myDB->isDBOpen() : false )
 			<< endl;
